Rename shared V2 format package to parquet.
Use a parquet-specific package name for DeltaParquetFileFormatV2 and ProtocolMetadataAdapterV2 to reflect their scope and address review naming feedback.
