+++ spark/v2/src/main/java/io/delta/spark/internal/v2/parquet/ProtocolMetadataAdapterV2.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.internal.v2.format;
+package io.delta.spark.internal.v2.parquet;
 
 import io.delta.kernel.internal.TableConfig;
 import io.delta.kernel.internal.actions.Metadata;
